refactor REST protocols - pull header generation code up from rest_xml into common module, execute it in rest_json protocol codegen as well
