+++ service_crategen/src/commands/generate/codegen/mod.rs
@@ -22,6 +22,7 @@ pub mod tests;
 mod rest_xml;
 mod xml_payload_parser;
 mod rest_response_parser;
+mod rest_request_generator;
 mod type_filter;
 
 type FileWriter = BufWriter<File>;
